[#818] better decoupling from standalone and eclipse astparserfactory
